pyproject.toml
[tool.poetry]
name = "observer"
version = "0.1.0"
description = ""
authors = ["Col Perks <wild.fauve@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pendulum = "^2.1.2"
rdflib = "^6.2.0"
PyMonad = "^2.4.0"
pyspark = "^3.3.0"
delta-spark = "^2.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-env = "^0.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
